gnu: jami: Always include the custom Jami patches with the sources.
Before this change, there was a Jami origin containing the sources with the
patches, and another one containing the sources but with the patches removed.
Unify the source to always contain the patches, which adds just 1 MiB to the
corresponding tarball and simplifies things.

* gnu/packages/jami.scm (%jami-sources): Make it a simple variable.  Move the
docstring to a comment.  Only remove the tarballs, always keeping the patches
with the sources.  Adjust the comments.
(%sfl-patches): Remove variable.
(jami-apply-dependency-patches): Rename to...
(jami-apply-custom-patches): ... this.  Adjust the input label used to
'jami-sources'.
(pjproject-jami): Replace %sfl-patches by %jami-sources.  Adjust following the
renaming of the jami-apply-custom-patches procedure.
(ffmpeg-jami): Likewise.  Do not make a temporary binding, simply call the
unquoted jami-apply-custom-patches procedure.