gnu: texlive-bin: Use ‘--with-banner-add’ configure option
The “Building TeX Live” manual has a section about configure options for
distro builds¹. They mention this option:

    --with-banner-add=/SomeDistro

    This isn’t technically required, but is strongly recommended, so your
    build and your distro can be distinguished from others.

Therefore add it to texlive-bin. This is how it appears on LuaTeX, for
example (it also shows up in other engines’ banners):

    $ luatex
    This is LuaTeX, Version 1.13.0 (TeX Live 2021/GNU Guix)
    restricted system commands enabled.
    **

The other options mentioned in that section are already being used.

¹ http://www.tug.org/texlive/doc/tlbuild.html#Distro-builds

* gnu/packages/tex.scm (texlive-bin)[arguments]: Add ‘--with-banner-add’
option.

Signed-off-by: Ludovic Courtès <ludo@gnu.org>