gnu: txr: Fix license installation.
The install-license-files phase was installing LICENSE-CYG when it
shouldn't have, because that license applies only to TXR builds for
Microsoft Windows; and it was also ignoring the METALICENSE file.
However, TXR's Makefile was already installing the LICENSE and
METALICENSE files into share/txr (the datadir), so redirect them into
the appropriate doc directory and delete the install-license-files
phase.

* gnu/packages/lisp.scm (txr)[arguments]<#:phases>
  {fix-license-installation}: New phase.
  {install-license-files}: Delete phase.

Signed-off-by: Guillaume Le Vaillant <glv@posteo.net>