gnu: weston: Fix package inputs.
The newly added inputs are referenced by the package but was missing in
the package definition. Some of these inputs might have been used
by the package indirectly via propagation from other inputs. The moved
inputs are referenced in pkg-config files.

* gnu/packages/freedesktop.scm (weston)[inputs]: Add glib, libdrm,
libpng, libx11 and libxcb. Move libxkbcommon and wayland to ...
[propagated-inputs]: ... here. Add pixman.