gnu: tryton modules: Unify native-inputs.
For testing, the tryton modules require a common set of native-input, which
basically are the propagated-inputs of `trytond`. (Some modules leave out one
or two of these, but most modules require all.) So to avoid redundant code,
use a variable for defining this common set. This reduces the maintenance
burden when more trytond modules are added.

* gnu/packages/tryton.scm (%standard-trytond-native-inputs): New variable.
  (trytond-account, trytond-account-invoice, trytond-account-invoice-stock,
  trytond-account-product, trytond-analytic-account, trytond-company,
  trytond-country, trytond-currency, trytond-party, trytond-product,
  trytond-purchase, trytond-purchase-request, trytond-stock, trytond-stock-lot,
  trytond-stock-supply)[native-inputs]: Use it.

native-inouts2