gnu: diffutils: Fix signal processing.
diffutils has a race condition in its signal processing code which is easy to
trigger on powerpc64le-linux.  More often than not, it causes the ‘colors’
test to fail and therefore the build of the package fails as well.

Add the patch proposed in Debian bug 922552 which fixes the problem.

* gnu/packages/patches/diffutils-fix-signal-processing.patch: New file.
* gnu/local.mk (dist_patch_DATA): Add it.
* gnu/packages/base.scm (diffutils)[source]: Use it.

Signed-off-by: Maxim Cournoyer <maxim.cournoyer@gmail.com>