gnu: gtk+: Fix testsuite failure on non-x86_64 platforms.
On non-x86_64 platforms, gtk+ is linked with a gdk-pixbuf version that
doesn’t support SVG.  This causes the ‘icontheme’ test to fail because it
tries to fetch SVG icons in some of the tests.

Therefore, we need to disable the test when SVG support isn’t available.

* gnu/packages/gtk.scm (gtk+)[propagated-inputs]: Use the ‘target-x86-64?’
convenience function.
[arguments]<#:phases>{disable-failing-tests}: Disable ‘icontheme’ test on
non-x86_64 platforms.

Signed-off-by: Mathieu Othacehe <othacehe@gnu.org>