gnu: xygrib: Fix finding resources in the data directory.
Currently, if you start up XyGrib it will not show icons in the toolbar
buttons, translated strings in the UI nor the map in the main view.  These
are all items stored in the data directory, which XyGrib can’t find.

Backport upstream PR 235 (already merged upstream) which makes it use XDG
directories correctly.

As a bonus, the PR also contains a fix to find libjpeg, so we don’t need
the substitution for that anymore.

* gnu/packages/patches/xygrib-fix-finding-data.patch: New file.
* gnu/local.mk (dist_patch_DATA): Add it.
* gnu/packages/geo.scm (xygrib)[source]: Use it.
  [arguments]<#:configure-flags>: Pass ‘-DGNU_PACKAGE=ON’.
  [arguments]<#:phases>{patch-directories}: Remove fix to find libjpeg.
  Remove trailing ‘#t’.
[native-search-paths]: Add ‘XDG_DATA_DIRS’.

Signed-off-by: Guillaume Le Vaillant <glv@posteo.net>