gnu: Rename google-brotli back to brotli.
Sorry for the noise, but good riddance.  Our nonstandard name for this
package has caused confusion since it was added.  For example, the
duplicate python-brotli and python-google-brotli packages added later,
with the latter name making even less sense.

The Python packages will be fixed in the next commit.

* gnu/packages/compression.scm (google-brotli): Redefine as a
deprecated-package, with…
(brotli): …restored to canonical status.
(python-google-brotli)[inherit]: Adjust accordingly.
* gnu/packages/databases.scm (apache-arrow)[inputs]: Adjust accordingly.
* gnu/packages/fontutils.scm (woff2)[inputs]: Likewise.
* gnu/packages/gnome.scm (libsoup)[propagated-inputs]: Likewise.
* gnu/packages/networking (wireshark)[inputs]: Likewise.
* gnu/packages/node.scm (node-lts)[arguments, native-inputs]: Likewise.
* gnu/packages/python-xyz.scm (python-imagecodecs)[inputs]: Likewise.