gnu: xpra: Install Cython-built libraries.
xpra would fail to start, because shared objects were
not installed. It’s not clear why it stopped working,
but it seems it was not the upgrade I reverted in commit
ab6d56d01d62757caa66fe0a23e8864c9122f0e6 to fix #52869. Sorry for the
confusion :(

* gnu/packages/patches/xpra-4.2-install_libs.patch: New file.
* gnu/local.mk (dist_patch_DATA): Register it.
* gnu/packages/xorg.scm (xpra)[source]: Add new patch.
[arguments]: Ensure 'build phase runs and passes #:configure-flags
to setup.py.