gnu: python-dbusmock: Update to 0.25.0 and reap terminated processes.
It was reported by abrenon and zimoun in #guix that python-dbusmock would fail
to build on their systems.

* gnu/packages/python-xyz.scm (python-dbusmock)
[imported-modules, modules]: New arguments.
[native-inputs]: Use new style.
[phases]{check}: Override phase to run it via a dumb init that can handle
signals and reap dead processes correctly.  This prevents test suite failures
in some cases (though it is still unknown why it was working on some systems
and not on others before the change).